refactor(cli): extract server AI and dashboard routes
