+++ packages/cli/src/server.ts
@@ -1,14 +1,6 @@
 import { createHash, randomUUID } from "node:crypto";
 import { type FSWatcher, watch as fsWatch } from "node:fs";
-import {
-  mkdir,
-  open as fsOpen,
-  readdir,
-  readFile,
-  stat,
-  unlink,
-  writeFile,
-} from "node:fs/promises";
+import { mkdir, open as fsOpen, readdir, readFile, stat, writeFile } from "node:fs/promises";
 import { homedir } from "node:os";
 import { join, resolve } from "node:path";
 import { serve } from "@hono/node-server";
@@ -22,18 +14,7 @@ import { classifyProject } from "@vibe-replay/types";
 import { readFileCache, writeFileCache } from "./cache.js";
 import { cleanPromptText, previewPrompt } from "./clean-prompt.js";
 import { computeDaysUntilCleanup, getClaudeCodeCleanupPeriod } from "./cleanup-warning.js";
-import {
-  type AiSelection,
-  generateFeedback,
-  generateToneAdjustment,
-  generateTranslation,
-} from "./feedback.js";
-import {
-  createBrowserAuthInteraction,
-  getAiRuntime,
-  readPiDefaultAiSelection,
-  type AiProviderInfo,
-} from "./ai-runtime.js";
+import { getAiRuntime } from "./ai-runtime.js";
 import {
   createLocalAssistantTools,
   type LocalAssistantAction,
@@ -41,13 +22,9 @@ import {
   type LocalAssistantContext,
   type LocalAssistantToolDetails,
 } from "./local-assistant.js";
-import { generateGitHubGif } from "./formatters/gif.js";
-import { generateGitHubMarkdown, generateGitHubSvg } from "./formatters/github.js";
 import { generateOutput, injectDataScript, loadViewerHtml } from "./generator.js";
-import { buildInsightsRollup } from "./insights-rollup.js";
 import { mergeInsights, readInsightsStore, writeInsightsStore } from "./insights.js";
-import { loadOverlays, sessionForExternalOutput, sessionWithEffectiveContent } from "./overlays.js";
-import { buildUsageCoverageReport } from "./usage-coverage.js";
+import { loadOverlays, sessionForExternalOutput } from "./overlays.js";
 import { parseClaudeCodeLines } from "./providers/claude-code/parser.js";
 import { parseCodexLines } from "./providers/codex/parser.js";
 import { parsePiLines } from "./providers/pi/parser.js";
@@ -57,14 +34,8 @@ import {
 } from "./providers/cursor/sqlite-reader.js";
 import { getAllProviders, getProvider } from "./providers/index.js";
 import { discoverProvidersSafely, type SafeProviderDiscoveryResult } from "./provider-discovery.js";
-import { getApiUrl, loadSavedCloudInfo, publishCloudWithOverlays } from "./publishers/cloud.js";
-import {
-  checkPublishStatus,
-  loadSavedGistInfo,
-  publishGist,
-  type SavedGistInfo,
-} from "./publishers/gist.js";
-import { scanForSecrets } from "./scan.js";
+import { getApiUrl, loadSavedCloudInfo } from "./publishers/cloud.js";
+import { loadSavedGistInfo, type SavedGistInfo } from "./publishers/gist.js";
 import { mergeSameSessions } from "./session-merge.js";
 import {
   getRemoteHome,
@@ -105,20 +76,21 @@ import type {
   SourceSessionCatalogCache,
   SourceSummaryRecord,
 } from "./server-types.js";
-import { loadAnnotations, saveAnnotations, saveOverlays } from "./server-persistence.js";
+import { loadAnnotations } from "./server-persistence.js";
 import { createAuthSession } from "./server-auth.js";
+import { resolveAiSelection, resolveDefaultAiSelection } from "./server-ai-selection.js";
 import { getArchivedSlugs, registerArchiveRoutes } from "./server-routes/archive.js";
 import { registerAuthRoutes } from "./server-routes/auth.js";
+import { registerAiRoutes } from "./server-routes/ai.js";
 import { registerCloudProxyRoutes } from "./server-routes/cloud.js";
+import { registerGenerationRoutes } from "./server-routes/generation.js";
+import { registerInsightsRoutes } from "./server-routes/insights.js";
 import { registerSessionAssetRoutes } from "./server-routes/session-assets.js";
+import { registerSessionOutputRoutes } from "./server-routes/session-output.js";
 import {
   buildInsightsSyncBatches,
   getErrorMessage,
-  hasReplayableContent,
-  type GenerateRequestBody,
-  replayOutputSlug,
   requireSlug,
-  resolveGenerateInputs,
   safeSlug,
   safeTargetId,
 } from "./server-core.js";
@@ -139,7 +111,7 @@ import {
   type UserInsights,
 } from "./scanner.js";
 import { transformToReplay } from "./transform.js";
-import type { ParsedTurn, ReplaySession, SessionInfo, SessionOverlays } from "./types.js";
+import type { ParsedTurn, ReplaySession, SessionInfo } from "./types.js";
 import { localDayKey, normalizeTitle } from "./utils.js";
 import { CLI_VERSION } from "./version.js";
 
@@ -213,131 +185,6 @@ function equivalentLocalOrigin(left: URL, right: URL): boolean {
     left.hostname === right.hostname ||
     (isLoopbackHostname(left.hostname) && isLoopbackHostname(right.hostname))
   );
-}
-
-function parseAiSelectionBody(body: unknown): AiSelection | undefined {
-  if (!body || typeof body !== "object" || Array.isArray(body)) return undefined;
-  const value = body as {
-    providerId?: unknown;
-    modelId?: unknown;
-    toolName?: unknown;
-  };
-  const providerId = value.providerId ?? value.toolName;
-  const modelId = value.modelId;
-  if (providerId === undefined && modelId === undefined) return undefined;
-  if (typeof providerId !== "string" || !providerId.trim()) {
-    throw new Error("providerId is required");
-  }
-  if (modelId !== undefined && typeof modelId !== "string") {
-    throw new Error("modelId must be a string");
-  }
-  return {
-    providerId: providerId.trim(),
-    ...(typeof modelId === "string" && modelId.trim() ? { modelId: modelId.trim() } : {}),
-  };
-}
-
-function normalizeAiBaseUrl(value: string): string {
-  try {
-    const url = new URL(value);
-    return `${url.protocol}//${url.host}${url.pathname.replace(/\/+$/, "")}`;
-  } catch {
-    return value.replace(/\/+$/, "");
-  }
-}
-
-async function resolveDefaultAiSelection(
-  providers: AiProviderInfo[],
-): Promise<AiSelection | undefined> {
-  const usable = providers.filter((provider) => provider.configured && provider.models.length > 0);
-  if (usable.length === 0) return undefined;
-
-  const piDefault = await readPiDefaultAiSelection();
-  if (piDefault?.providerId) {
-    const matchingProvider = usable.find((provider) => provider.id === piDefault.providerId);
-    if (
-      matchingProvider &&
-      piDefault.modelId &&
-      matchingProvider.models.some((model) => model.id === piDefault.modelId)
-    ) {
-      return {
-        providerId: matchingProvider.id,
-        modelId: piDefault.modelId,
-      };
-    }
-  }
-
-  // Provider ids are not globally stable across runtimes. For custom
-  // providers, use the configured endpoint as the identity and only accept
-  // the default model when that exact provider exposes it.
-  if (piDefault?.baseUrl && piDefault.modelId) {
-    const normalizedBaseUrl = normalizeAiBaseUrl(piDefault.baseUrl);
-    const matchingProvider = usable.find(
-      (provider) =>
-        provider.custom &&
-        normalizeAiBaseUrl(provider.custom.baseUrl) === normalizedBaseUrl &&
-        provider.models.some((model) => model.id === piDefault.modelId),
-    );
-    if (matchingProvider) {
-      return { providerId: matchingProvider.id, modelId: piDefault.modelId };
-    }
-  }
-
-  // Do not invent a provider/model when there is no explicit cross-runtime
-  // identity. The UI should ask the user to choose one instead of silently
-  // selecting the first catalog entry.
-  return undefined;
-}
-
-async function resolveAiSelection(
-  body: unknown,
-  signal?: AbortSignal,
-): Promise<{
-  selection: AiSelection;
-  providerName: string;
-  modelId: string;
-  authType: string;
-  authSubscription: boolean;
-  authSource?: string;
-}> {
-  const runtime = getAiRuntime();
-  const requested = parseAiSelectionBody(body);
-  let providerId = requested?.providerId;
-  let defaultModelId: string | undefined;
-  if (!providerId) {
-    const providers = await runtime.listProviders({ signal });
-    const defaultSelection = await resolveDefaultAiSelection(providers);
-    providerId = defaultSelection?.providerId;
-    defaultModelId = defaultSelection?.modelId;
-    if (!providerId) {
-      const hasUsableProvider = providers.some(
-        (provider) => provider.configured && provider.models.length > 0,
-      );
-      throw new Error(
-        hasUsableProvider
-          ? "No AI provider/model is selected. Choose a provider and model in AI Studio."
-          : "No usable AI provider is configured. Set up a provider and model in AI Studio.",
-      );
-    }
-  }
-
-  const modelId = requested?.modelId ?? defaultModelId;
-  if (!modelId) {
-    throw new Error("No AI model is selected. Choose a provider and model in AI Studio.");
-  }
-  const resolved = await runtime.resolveModel(providerId, modelId, { signal });
-  return {
-    selection: {
-      providerId: resolved.provider.id,
-      modelId: resolved.model.id,
-    },
-    providerName: resolved.provider.name,
-    modelId: resolved.model.id,
-    authType: resolved.auth.type,
-    authSubscription:
-      resolved.auth.type === "oauth" && resolved.provider.auth.oauth?.isSubscription === true,
-    authSource: resolved.auth.source,
-  };
 }
 
 const LOCAL_ASSISTANT_SYSTEM_PROMPT = `You are the Vibe Replay local assistant.
@@ -2854,413 +2701,21 @@ export async function startServer(
   app.get("/api/sources/stream", streamSourceSessions);
   app.get("/api/source-sessions/stream", streamSourceSessions);
 
-  // --- Generate: parse a source session into a replay ---
-  app.post("/api/generate", async (c) => {
-    try {
-      const body = await c.req.json<GenerateRequestBody>();
-
-      const provider = getProvider(body.provider);
-      if (!provider) {
-        return c.json({ error: `Unknown provider: ${body.provider}` }, 400);
-      }
-
-      let discoveredSessions: SessionInfo[] = [];
-      if (typeof body.sessionSlug === "string" && safeSlug(body.sessionSlug)) {
-        discoveredSessions = mergeSameSessions(
-          (await discoverProvidersSafely([provider])).sessions,
-        );
-      }
-
-      const resolved = resolveGenerateInputs(body, discoveredSessions);
-      if (!resolved.ok) {
-        return c.json({ error: resolved.error }, 400);
-      }
-      if (body.title !== undefined && typeof body.title !== "string") {
-        return c.json({ error: "title must be a string" }, 400);
-      }
-
-      const parsed = await provider.parse(resolved.value.paths, resolved.value.sessionInfo);
-
-      const home = homedir();
-      const rawProject = body.sessionProject || parsed.cwd;
-      const project = shortenPath(rawProject, home);
-      const gitRepo =
-        resolved.value.sessionInfo?.gitRepo ||
-        (resolved.value.sessionInfo?.location?.kind === "ssh"
-          ? undefined
-          : await readGitRepo(rawProject));
-
-      const replay = transformToReplay(parsed, body.provider, project, {
-        generator: {
-          name: "vibe-replay",
-          version: CLI_VERSION,
-          generatedAt: new Date().toISOString(),
-        },
-        gitRepo,
-        location: resolved.value.sessionInfo?.location,
-        remoteHome: getRemoteHome(
-          resolved.value.sessionInfo?.location?.kind === "ssh"
-            ? resolved.value.sessionInfo.location.id
-            : undefined,
-        ),
-      });
-      if (!hasReplayableContent(replay)) {
-        return c.json({ error: "This session has no replayable user prompts" }, 422);
-      }
-
-      if (typeof body.title === "string") {
-        const normalizedCustomTitle = normalizeTitle(body.title);
-        if (normalizedCustomTitle) {
-          replay.meta.title = normalizedCustomTitle;
-        }
-      }
-
-      // Save replay
-      const rawSlug = replay.meta.slug || replay.meta.sessionId.slice(0, 8);
-      const slug = replayOutputSlug(rawSlug, resolved.value.sessionInfo?.location, {
-        provider: body.provider,
-        sessionId: replay.meta.sessionId,
-      });
-      const outputDir = join(baseDir, slug);
-      await generateOutput(replay, outputDir);
-      const updatedReplays = await refreshReplaysCache();
-      if (updatedReplays) await syncSourcesCacheWithReplays(updatedReplays);
-
-      // Secret scanning
-      const findings = scanForSecrets(JSON.stringify(replay));
-      const warnings = findings.map((f) => `[${f.rule}] ${f.match}`);
-
-      return c.json({
-        slug,
-        title: replay.meta.title || slug,
-        sceneCount: replay.scenes.length,
-        stats: {
-          userPrompts: replay.meta.stats.userPrompts,
-          toolCalls: replay.meta.stats.toolCalls,
-          thinkingBlocks: replay.meta.stats.thinkingBlocks,
-        },
-        warnings: warnings.length > 0 ? warnings : undefined,
-      });
-    } catch (err) {
-      return c.json({ error: getErrorMessage(err) }, 500);
-    }
+  registerGenerationRoutes(app, {
+    baseDir,
+    discoverAllProviders: () => discoverAllProviders(),
+    refreshReplaysCache,
+    syncSourcesCacheWithReplays,
   });
 
-  // --- Regenerate all existing replays ---
-  app.post("/api/regenerate-all", async (c) => {
-    const replaysDir = baseDir;
-    const { readdir, readFile: readF } = await import("node:fs/promises");
-    const results: Array<{ slug: string; status: string; scenes?: number }> = [];
-
-    // Discover all sessions across all providers and configured SSH sources.
-    const allProviders = getAllProviders();
-    const allSessions = mergeSameSessions((await discoverAllProviders()).sessions);
-
-    let entries: string[];
-    try {
-      entries = await readdir(replaysDir);
-    } catch {
-      return c.json({ error: "No replays directory" }, 404);
-    }
-
-    for (const slug of entries) {
-      if (slug.startsWith(".") || slug === "cache") continue;
-      try {
-        const replayPath = join(replaysDir, slug, "replay.json");
-        const raw = await readF(replayPath, "utf-8").catch(() => null);
-        if (!raw) continue;
-
-        const oldReplay = JSON.parse(raw);
-        const sessionId = oldReplay.meta?.sessionId;
-        const providerName = oldReplay.meta?.provider || "claude-code";
-        if (!sessionId) {
-          results.push({ slug, status: "skipped: no sessionId" });
-          continue;
-        }
-
-        // Find source session by sessionId
-        const replayTargetId =
-          oldReplay.meta?.location?.kind === "ssh" && typeof oldReplay.meta.location.id === "string"
-            ? oldReplay.meta.location.id
-            : undefined;
-        const sessionInfo = allSessions.find((s) => {
-          const sessionTargetId = s.location?.kind === "ssh" ? s.location.id : undefined;
-          return (
-            s.provider === providerName &&
-            s.sessionId === sessionId &&
-            sessionTargetId === replayTargetId
-          );
-        });
-        if (!sessionInfo || sessionInfo.filePaths.length === 0) {
-          results.push({
-            slug,
-            status: sessionInfo?.transcriptStatus
-              ? `skipped: ${sessionInfo.transcriptStatus}`
-              : "skipped: source not found",
-          });
-          continue;
-        }
-        if (sessionInfo.transcriptStatus) {
-          results.push({ slug, status: `skipped: ${sessionInfo.transcriptStatus}` });
-          continue;
-        }
-
-        const provider = allProviders.find((p) => p.name === providerName);
-        if (!provider) {
-          results.push({ slug, status: `skipped: unknown provider ${providerName}` });
-          continue;
-        }
-
-        // Re-parse and re-generate
-        const paths = [...sessionInfo.filePaths, ...(sessionInfo.toolPaths || [])];
-        const parsed = await provider.parse(paths, sessionInfo);
-        const home = homedir();
-        const project = shortenPath(sessionInfo.project, home);
-
-        const replay = transformToReplay(parsed, providerName, project, {
-          generator: {
-            name: "vibe-replay",
-            version: CLI_VERSION,
-            generatedAt: new Date().toISOString(),
-          },
-          gitRepo: sessionInfo.gitRepo,
-          location: sessionInfo.location,
-          transcriptStatus: sessionInfo.transcriptStatus,
-          remoteHome: getRemoteHome(
-            sessionInfo.location?.kind === "ssh" ? sessionInfo.location.id : undefined,
-          ),
-        });
-        if (!hasReplayableContent(replay)) {
-          results.push({ slug, status: "skipped: no replayable user prompts" });
-          continue;
-        }
-
-        // Preserve custom title from old replay
-        if (oldReplay.meta?.title) replay.meta.title = oldReplay.meta.title;
-
-        const outputDir = join(replaysDir, slug);
-        await generateOutput(replay, outputDir);
-        results.push({ slug, status: "regenerated", scenes: replay.scenes.length });
-      } catch (err) {
-        results.push({ slug, status: `error: ${getErrorMessage(err)}` });
-      }
-    }
-
-    const updatedReplays = await refreshReplaysCache();
-    if (updatedReplays) await syncSourcesCacheWithReplays(updatedReplays);
-    return c.json({
-      total: results.length,
-      regenerated: results.filter((r) => r.status === "regenerated").length,
-      results,
-    });
-  });
-
-  // --- Session scanner: background metadata extraction for insights ---
-
-  app.post("/api/scan/start", async (c) => {
-    const hints = enrichmentHintsFromBody(await c.req.json().catch(() => undefined));
-    const queued = requestBackgroundScan(hints);
-    return c.json({
-      ok: true,
-      queued,
-      message: queued ? "Background scan queued" : "Background scan started",
-    });
-  });
-
-  app.get("/api/scan/status", async (c) => {
-    return c.json({
-      running: scanState.running,
-      scanned: scanState.scanned,
-      total: scanState.total,
-      resultCount: scanState.results.length,
-      revision: scanState.revision,
-      hasSnapshot: scanState.hasSnapshot,
-      currentSession: scanState.currentSession,
-      phase: scanState.phase,
-      startedAt: scanState.startedAt,
-      finishedAt: scanState.finishedAt,
-      failedProviders: scanState.failedProviders || [],
-      usageBackfill: scanState.usageBackfill,
-      usageIndexPending: countPendingCursorUsageIndexes(scanState.results),
-      hasInsights: insightsCache.userInsights !== null,
-      hasCachedResults: scanState.results.length > 0,
-      cachedResultCount: scanState.results.length,
-      cachedAt: scanState.finishedAt,
-    });
-  });
-
-  app.get("/api/scan/results", async (c) => {
-    return c.json({
-      // Invocation events are available from the focused usage endpoint; keep
-      // the dashboard's initial scan payload bounded to per-session summaries.
-      results: scanState.results.map((result) => ({ ...result, usageEvents: undefined })),
-      running: scanState.running,
-      scanned: scanState.scanned,
-      total: scanState.total,
-      revision: scanState.revision,
-      finishedAt: scanState.finishedAt,
-      failedProviders: scanState.failedProviders || [],
-    });
-  });
-
-  app.get("/api/usage/events", async (c) => {
-    const provider = c.req.query("provider");
-    const sessionId = c.req.query("sessionId");
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    if (!provider || !sessionId) {
-      return c.json({ error: "provider and sessionId are required" }, 400);
-    }
-    const result = scanState.results.find(
-      (scan) =>
-        scan.provider === provider &&
-        scan.sessionId === sessionId &&
-        (targetId
-          ? scan.location?.kind === "ssh" && scan.location.id === targetId
-          : scan.location?.kind !== "ssh"),
-    );
-    if (!result) return c.json({ error: "Session usage not found" }, 404);
-    return c.json({
-      provider,
-      sessionId,
-      ...(result.location ? { location: result.location } : {}),
-      summary: result.usageSummary,
-      events: result.usageEvents || [],
-    });
-  });
-
-  // Compact projection of the scan results: just enough to aggregate usage over
-  // any time range client-side, without shipping the multi-MB scan payload.
-  app.get("/api/usage/rollup", async (c) => {
-    const sessions = scanState.results
-      .filter((scan) => scan.usageSummary)
-      .map((scan) => ({
-        provider: scan.provider,
-        sessionId: scan.sessionId,
-        ...(scan.location ? { location: scan.location } : {}),
-        project: scan.project,
-        startTime: scan.startTime,
-        usage: scan.usageSummary,
-      }));
-    return c.json({
-      sessions,
-      indexedSessions: scanState.results.filter((scan) => scan.usageIndexed === true).length,
-      totalSessions: scanState.results.length,
-      scannedAt: scanState.finishedAt,
-      coverage: buildUsageCoverageReport(scanState.results),
-    });
-  });
-
-  // Compact per-session projection for exact 7d/30d/90d insight totals and
-  // range-scoped secondary breakdowns. Conversation content and tool
-  // inputs/results intentionally never leave the scanner here.
-  app.get("/api/insights/rollup", async (c) => {
-    if (!scanState.hasSnapshot) {
-      return c.json({ error: "No scan results available. Start a scan first." }, 503);
-    }
-    // Replay files can be created or deleted without a source scan, so refresh
-    // this small list instead of trusting a potentially stale dashboard cache.
-    const refreshedReplays = await refreshReplaysCache();
-    let replays = refreshedReplays;
-    if (!replays) {
-      const cachedReplays = await readFileCache<ReplaySummary[]>(replaysCacheKey);
-      replays = cachedReplays?.data || [];
-    }
-    return c.json(buildInsightsRollup(scanState.results, replays, { includeDetails: true }));
-  });
-
-  app.get("/api/insights", async (c) => {
-    const project = c.req.query("project");
-
-    if (project) {
-      const targetId = safeTargetId(c.req.query("targetId"));
-      if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-      const location: ProjectInsightLocation = targetId
-        ? (scanState.results.find(
-            (scan) => scan.location?.kind === "ssh" && scan.location.id === targetId,
-          )?.location ?? { kind: "ssh", id: targetId, label: targetId })
-        : "local";
-      const cacheKey = projectInsightKey(project, undefined, location);
-      // Project-level: cache hit → O(1), miss → compute on demand
-      const cached =
-        insightsCache.projectInsights.get(cacheKey) ||
-        (location === "local" ? insightsCache.projectInsights.get(project) : undefined);
-      if (cached) return c.json({ type: "project", insights: cached });
-
-      // Fallback: compute on demand
-      const scans = scanState.results;
-      if (!scans.length) {
-        return c.json({ error: "No scan results available. Start a scan first." }, 404);
-      }
-      const memory = location === "local" ? await readProjectMemory(project) : undefined;
-      const insights = aggregateProjectInsights(project, scans, memory || undefined, location);
-      insightsCache.projectInsights.set(cacheKey, insights);
-      return c.json({ type: "project", insights });
-    }
-
-    // User-level: cache hit → O(1)
-    if (insightsCache.userInsights) {
-      return c.json({ type: "user", insights: insightsCache.userInsights });
-    }
-
-    // Fallback: compute on demand
-    const scans = scanState.results;
-    if (!scans.length) {
-      return c.json({ error: "No scan results available. Start a scan first." }, 404);
-    }
-    const insights = aggregateUserInsights(scans);
-    return c.json({ type: "user", insights });
-  });
-
-  // --- Local insights store (durable, survives source deletion) ---
-
-  app.get("/api/insights/local", async (c) => {
-    const store = await readInsightsStore();
-    return c.json(store);
-  });
-
-  app.get("/api/insights/local/stats", async (c) => {
-    const { getInsightsStats } = await import("./insights.js");
-    const store = await readInsightsStore();
-    return c.json(getInsightsStats(store));
-  });
-
-  app.post("/api/insights/sync", async (c) => {
-    const result = await syncInsightsToCloud();
-    if (result.error === "Not logged in") {
-      return c.json({ error: "Not logged in. Run: vibe-replay auth login" }, 401);
-    }
-    if (result.error) {
-      return c.json({
-        error: `Sync failed: ${result.error}`,
-        synced: result.synced,
-        total: result.total,
-      });
-    }
-    if (result.total === 0) {
-      return c.json({ synced: 0, message: "All insights already synced" });
-    }
-    return c.json({
-      synced: result.synced,
-      total: result.total,
-      message: `Synced ${result.synced} insights to cloud`,
-    });
-  });
-
-  app.get("/api/memory", async (c) => {
-    const project = c.req.query("project");
-    if (!project) return c.json({ error: "project parameter required" }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    if (targetId !== undefined) {
-      // Project memory is read from the local filesystem. Never let a remote
-      // project request accidentally read a same-named local project's memory.
-      return c.json({ memoryFiles: [], claudeMd: null });
-    }
-
-    const memory = await readProjectMemory(project);
-    if (!memory) return c.json({ memoryFiles: [], claudeMd: null });
-    return c.json(memory);
+  registerInsightsRoutes(app, {
+    getScanState: () => scanState,
+    getInsightsCache: () => insightsCache,
+    requestBackgroundScan,
+    refreshReplaysCache,
+    readCachedReplays: async () =>
+      (await readFileCache<ReplaySummary[]>(replaysCacheKey))?.data || [],
+    syncInsightsToCloud,
   });
 
   registerSessionAssetRoutes(app, {
@@ -3268,9 +2723,9 @@ export async function startServer(
     loadSession: (slug, targetId) => loadSessionFromDisk(baseDir, slug, targetId),
   });
 
-  // GitHub CLI status
-  app.get("/api/gh-status", (c) => {
-    return c.json(checkPublishStatus());
+  registerSessionOutputRoutes(app, {
+    baseDir,
+    loadSession: (slug, targetId) => loadSessionFromDisk(baseDir, slug, targetId),
   });
 
   // Auth — read local auth.json (per-environment, keyed by API origin)
@@ -3332,649 +2787,10 @@ export async function startServer(
     }
   });
 
-  // Gist info for a session (requires slug)
-  app.get("/api/gist-info", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-    } catch {
-      return c.json({ error: "session not found" }, 404);
-    }
-    const targetDir = join(baseDir, result.slug);
-    const gist = await loadSavedGistInfo(targetDir);
-    if (!gist) return c.json({ gist: null });
-    return c.json({ gist });
-  });
-
-  // Delete stale gist info (gist deleted on GitHub)
-  app.delete("/api/gist-info", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-    } catch {
-      return c.json({ error: "session not found" }, 404);
-    }
-    const metaPath = join(baseDir, result.slug, ".vibe-replay-gist.json");
-    await unlink(metaPath).catch(() => {});
-    return c.json({ ok: true });
-  });
-
-  // Cloud info for a session (requires slug)
-  app.get("/api/cloud-info", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-    } catch {
-      return c.json({ error: "session not found" }, 404);
-    }
-    const targetDir = join(baseDir, result.slug);
-    const cloud = await loadSavedCloudInfo(targetDir);
-    if (!cloud) return c.json({ cloud: null });
-    return c.json({ cloud });
-  });
-
-  // Save cloud info locally (after browser-side upload)
-  app.post("/api/cloud-info", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-    } catch {
-      return c.json({ error: "session not found" }, 404);
-    }
-    const targetDir = join(baseDir, result.slug);
-    const body = await c.req.json();
-    if (!body.id || !body.url) return c.json({ error: "Missing id/url" }, 400);
-    const metaPath = join(targetDir, ".vibe-replay-cloud.json");
-    await writeFile(
-      metaPath,
-      JSON.stringify(
-        {
-          id: body.id,
-          url: body.url,
-          expiresAt: body.expiresAt,
-          updatedAt: new Date().toISOString(),
-        },
-        null,
-        2,
-      ),
-      "utf-8",
-    );
-    return c.json({ ok: true });
-  });
-
-  // Delete cloud info locally
-  app.delete("/api/cloud-info", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-    } catch {
-      return c.json({ error: "session not found" }, 404);
-    }
-    const metaPath = join(baseDir, result.slug, ".vibe-replay-cloud.json");
-    await unlink(metaPath).catch(() => {});
-    return c.json({ ok: true });
-  });
-
-  // Publish to Gist (requires slug)
-  app.post("/api/publish/gist", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    const targetDir = join(baseDir, result.slug);
-
-    try {
-      const rawSession = await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const overlaysData = await loadOverlays(baseDir, result.slug, targetId);
-      const targetSession = sessionForExternalOutput(
-        sessionWithEffectiveContent(rawSession, overlaysData),
-      );
-
-      // Write effective content for gist, then restore the original replay.json
-      const replayPath = join(targetDir, "replay.json");
-      const originalContent = await readFile(replayPath, "utf-8");
-      await writeFile(replayPath, JSON.stringify(targetSession), "utf-8");
-
-      try {
-        const title = targetSession.meta.title || targetSession.meta.slug;
-        const savedGist = await loadSavedGistInfo(targetDir);
-        const gistResult = await publishGist(targetDir, title, {
-          overwrite: savedGist || undefined,
-        });
-        return c.json(gistResult);
-      } finally {
-        // Always restore original replay.json
-        await writeFile(replayPath, originalContent, "utf-8");
-      }
-    } catch (err) {
-      return c.json({ error: getErrorMessage(err) }, 500);
-    }
-  });
-
-  // Publish to cloud (R2) — overlay merging is handled by publishCloudWithOverlays
-  app.post("/api/publish/cloud", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    const targetDir = join(baseDir, result.slug);
-
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const body = await c.req.json().catch(() => ({}));
-      const cloudResult = await publishCloudWithOverlays(targetDir, {
-        visibility: body.visibility || "unlisted",
-        targetId: targetId || undefined,
-      });
-      return c.json(cloudResult);
-    } catch (err) {
-      return c.json({ error: getErrorMessage(err) }, 500);
-    }
-  });
-
-  // Export HTML (requires slug)
-  app.post("/api/export/html", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    const targetDir = join(baseDir, result.slug);
-
-    try {
-      const rawSession = await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const overlaysData = await loadOverlays(baseDir, result.slug, targetId);
-      const targetSession = sessionForExternalOutput(
-        sessionWithEffectiveContent(rawSession, overlaysData),
-      );
-
-      // generateOutput writes replay.json — save/restore to avoid destructive overwrite
-      const replayPath = join(targetDir, "replay.json");
-      const originalContent = await readFile(replayPath, "utf-8");
-      try {
-        const outputPath = await generateOutput(targetSession, targetDir);
-        return c.json({ path: outputPath });
-      } finally {
-        await writeFile(replayPath, originalContent, "utf-8");
-      }
-    } catch (err) {
-      return c.json({ error: getErrorMessage(err) }, 500);
-    }
-  });
-
-  // Check existing GitHub export files (requires slug)
-  app.get("/api/export/github/status", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    const targetDir = join(baseDir, result.slug);
-    try {
-      await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const svgPath = join(targetDir, "session-preview.svg");
-      const mdPath = join(targetDir, "github-summary.md");
-      const gifPath = join(targetDir, "session-preview.gif");
-      const [svgContent, markdown, gifBuf] = await Promise.all([
-        readFile(svgPath, "utf-8").catch(() => null),
-        readFile(mdPath, "utf-8").catch(() => null),
-        readFile(gifPath).catch(() => null),
-      ]);
-      if (!svgContent && !markdown && !gifBuf) return c.json({ exists: false });
-      const gist = await loadSavedGistInfo(targetDir);
-      const gifContent = gifBuf ? gifBuf.toString("base64") : null;
-      // Get file modification times for "last generated" display
-      const [gifMtime, svgMtime, mdMtime] = await Promise.all([
-        stat(gifPath)
-          .then((s) => s.mtime.toISOString())
-          .catch(() => null),
-        stat(svgPath)
-          .then((s) => s.mtime.toISOString())
-          .catch(() => null),
-        stat(mdPath)
-          .then((s) => s.mtime.toISOString())
-          .catch(() => null),
-      ]);
-      return c.json({
-        exists: true,
-        svgContent,
-        markdown,
-        svgPath,
-        mdPath,
-        gifContent,
-        gifPath,
-        gifGeneratedAt: gifMtime,
-        svgGeneratedAt: svgMtime,
-        mdGeneratedAt: mdMtime,
-        replayUrl: gist?.viewerUrl || undefined,
-      });
-    } catch {
-      return c.json({ exists: false });
-    }
-  });
-
-  // Export GitHub markdown + SVG + GIF (requires slug)
-  app.post("/api/export/github", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-    const targetDir = join(baseDir, result.slug);
-
-    try {
-      const rawSession = await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const overlaysData = await loadOverlays(baseDir, result.slug, targetId);
-      const targetSession = sessionForExternalOutput(
-        sessionWithEffectiveContent(rawSession, overlaysData),
-      );
-
-      // Check for a previously published gist to use as replay URL
-      const gist = await loadSavedGistInfo(targetDir);
-      const replayUrl = gist?.viewerUrl || undefined;
-
-      // Generate SVG
-      const svgContent = generateGitHubSvg(targetSession, { replayUrl });
-      const svgFilePath = join(targetDir, "session-preview.svg");
-      await writeFile(svgFilePath, svgContent, "utf-8");
-
-      // Generate GIF
-      let gifContent: string | null = null;
-      let gifFilePath: string | null = null;
-      let gifWarning: string | undefined;
-      try {
-        const gifBuffer = await generateGitHubGif(targetSession, { replayUrl });
-        gifFilePath = join(targetDir, "session-preview.gif");
-        await writeFile(gifFilePath, gifBuffer);
-        gifContent = gifBuffer.toString("base64");
-      } catch (err) {
-        // GIF generation is best-effort — SVG still works
-        gifWarning = `GIF generation failed: ${err instanceof Error ? err.message : String(err)}`;
-      }
-
-      // Generate markdown (prefer GIF for universal GitHub support)
-      const markdown = generateGitHubMarkdown(targetSession, {
-        replayUrl,
-        svgPath: "./session-preview.svg",
-        gifPath: gifContent ? "./session-preview.gif" : undefined,
-      });
-      const mdFilePath = join(targetDir, "github-summary.md");
-      await writeFile(mdFilePath, markdown, "utf-8");
-
-      // Secret scan warnings
-      const findings = scanForSecrets(JSON.stringify(targetSession));
-      const warnings = findings.map((f) => `[${f.rule}] ${f.match}`);
-
-      const now = new Date().toISOString();
-      return c.json({
-        markdown,
-        svgContent,
-        svgPath: svgFilePath,
-        mdPath: mdFilePath,
-        gifContent,
-        gifPath: gifFilePath,
-        gifGeneratedAt: gifContent ? now : undefined,
-        gifWarning,
-        svgGeneratedAt: now,
-        mdGeneratedAt: now,
-        replayUrl,
-        warnings: warnings.length > 0 ? warnings : undefined,
-      });
-    } catch (err) {
-      return c.json({ error: getErrorMessage(err) }, 500);
-    }
-  });
-
-  // AI Studio — embedded Pi provider registry and authentication.
-  const getAiProvidersResponse = async (signal?: AbortSignal) => {
-    const providers = await getAiRuntime().listProviders({ signal });
-    const defaultSelection = await resolveDefaultAiSelection(providers);
-    const defaultProvider = defaultSelection
-      ? providers.find((provider) => provider.id === defaultSelection.providerId) || null
-      : null;
-    return {
-      available: providers.some((provider) => provider.configured),
-      providers,
-      defaultProvider: defaultProvider
-        ? {
-            id: defaultProvider.id,
-            name: defaultProvider.name,
-            ...(defaultSelection?.modelId ? { modelId: defaultSelection.modelId } : {}),
-          }
-        : null,
-    };
-  };
-
-  app.get("/api/ai/providers", async (c) => {
-    try {
-      return c.json(await getAiProvidersResponse(c.req.raw.signal));
-    } catch (err) {
-      return c.json(
-        { available: false, providers: [], error: await getAiRuntime().getSafeErrorMessage(err) },
-        500,
-      );
-    }
-  });
-
-  // Keep the old endpoint name as a compatibility alias for existing viewers.
-  app.get("/api/feedback/detect", async (c) => {
-    try {
-      const response = await getAiProvidersResponse(c.req.raw.signal);
-      const defaultProvider = response.defaultProvider;
-      return c.json({
-        ...response,
-        tool: defaultProvider ? { name: defaultProvider.id } : undefined,
-        tools: response.providers.map((provider) => ({
-          name: provider.id,
-          label: provider.name,
-        })),
-        defaultTool: defaultProvider ? { name: defaultProvider.id } : undefined,
-      });
-    } catch (err) {
-      return c.json(
-        { available: false, providers: [], error: await getAiRuntime().getSafeErrorMessage(err) },
-        500,
-      );
-    }
-  });
-
-  app.post("/api/ai/custom", async (c) => {
-    if (!isSameOriginSettingsRequest(c)) {
-      return c.json({ error: "AI provider requests must be same-origin" }, 403);
-    }
-
-    const body = await c.req.json().catch(() => undefined);
-    if (!body || typeof body !== "object" || Array.isArray(body)) {
-      return c.json({ error: "Request body must be an object" }, 400);
-    }
-    const value = body as {
-      name?: unknown;
-      baseUrl?: unknown;
-      apiKey?: unknown;
-    };
-    if (typeof value.baseUrl !== "string" || !value.baseUrl.trim()) {
-      return c.json({ error: "baseUrl is required" }, 400);
-    }
-    if (value.name !== undefined && typeof value.name !== "string") {
-      return c.json({ error: "name must be a string" }, 400);
-    }
-    if (value.apiKey !== undefined && typeof value.apiKey !== "string") {
-      return c.json({ error: "apiKey must be a string" }, 400);
-    }
-
-    try {
-      const runtime = getAiRuntime();
-      await runtime.configureCustomProvider(
-        {
-          baseUrl: value.baseUrl,
-          ...(typeof value.name === "string" ? { name: value.name } : {}),
-        },
-        typeof value.apiKey === "string" ? value.apiKey : undefined,
-        c.req.raw.signal,
-      );
-      return c.json({ ok: true, ...(await getAiProvidersResponse(c.req.raw.signal)) });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 400);
-    }
-  });
-
-  app.delete("/api/ai/custom", async (c) => {
-    if (!isSameOriginSettingsRequest(c)) {
-      return c.json({ error: "AI provider requests must be same-origin" }, 403);
-    }
-    try {
-      const runtime = getAiRuntime();
-      await runtime.removeCustomProvider(c.req.raw.signal);
-      return c.json({ ok: true, ...(await getAiProvidersResponse(c.req.raw.signal)) });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 400);
-    }
-  });
-
-  app.post("/api/ai/auth", async (c) => {
-    if (!isSameOriginSettingsRequest(c)) {
-      return c.json({ error: "AI authentication requests must be same-origin" }, 403);
-    }
-
-    const body = await c.req.json().catch(() => undefined);
-    if (!body || typeof body !== "object" || Array.isArray(body)) {
-      return c.json({ error: "Request body must be an object" }, 400);
-    }
-
-    const value = body as {
-      providerId?: unknown;
-      method?: unknown;
-      apiKey?: unknown;
-    };
-    if (typeof value.providerId !== "string" || !value.providerId.trim()) {
-      return c.json({ error: "providerId is required" }, 400);
-    }
-    if (value.method !== "api_key" && value.method !== "oauth") {
-      return c.json({ error: "method must be api_key or oauth" }, 400);
-    }
-
-    try {
-      const runtime = getAiRuntime();
-      const providerId = value.providerId.trim();
-      if (value.method === "api_key") {
-        if (typeof value.apiKey !== "string") {
-          return c.json({ error: "apiKey is required for API-key authentication" }, 400);
-        }
-        await runtime.saveApiKey(providerId, value.apiKey, c.req.raw.signal);
-      } else {
-        await runtime.login(providerId, "oauth", createBrowserAuthInteraction(c.req.raw.signal));
-      }
-      return c.json({ ok: true, ...(await getAiProvidersResponse(c.req.raw.signal)) });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 400);
-    }
-  });
-
-  app.delete("/api/ai/auth", async (c) => {
-    if (!isSameOriginSettingsRequest(c)) {
-      return c.json({ error: "AI authentication requests must be same-origin" }, 403);
-    }
-    const providerId = c.req.query("providerId")?.trim();
-    if (!providerId) return c.json({ error: "providerId is required" }, 400);
-    try {
-      const runtime = getAiRuntime();
-      await runtime.logout(providerId, c.req.raw.signal);
-      return c.json({ ok: true, ...(await getAiProvidersResponse(c.req.raw.signal)) });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 400);
-    }
-  });
-
-  // AI Feedback — generate feedback annotations (requires slug)
-  app.post("/api/feedback/generate", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-
-    try {
-      const body = await c.req.json().catch(() => ({}));
-      const ai = await resolveAiSelection(body, c.req.raw.signal);
-
-      const targetSession = await loadSessionFromDisk(baseDir, result.slug, targetId);
-
-      const fb = await generateFeedback(targetSession, ai.selection, {
-        signal: c.req.raw.signal,
-      });
-      if (!fb) return c.json({ error: "AI Coach returned no feedback" }, 422);
-
-      const existingAnns = targetSession.annotations ?? [];
-      const newAnnotations = [
-        ...existingAnns.filter((a) => a.author !== "vibe-feedback"),
-        ...fb.annotations,
-      ];
-
-      // Persist
-      try {
-        await saveAnnotations(baseDir, result.slug, newAnnotations, targetId);
-      } catch {
-        /* ignore */
-      }
-
-      return c.json({
-        annotations: newAnnotations,
-        score: fb.result.score,
-        itemCount: fb.result.feedbackItems.length,
-        outcome: fb.result.outcome,
-        sessionGoal: fb.result.sessionGoal,
-        providerId: ai.selection.providerId,
-        providerName: ai.providerName,
-        modelId: ai.modelId,
-        authType: ai.authType,
-        authSubscription: ai.authSubscription,
-        authSource: ai.authSource,
-      });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 500);
-    }
-  });
-
-  // After generation, fix originalValue to be the TRUE original from the unmodified session
-  function fixOriginalValues(
-    overlays: import("./types.js").SceneOverlay[],
-    originalSession: ReplaySession,
-  ) {
-    for (const overlay of overlays) {
-      const scene = originalSession.scenes[overlay.sceneIndex];
-      if (scene && (scene.type === "user-prompt" || scene.type === "text-response")) {
-        overlay.originalValue = scene.content;
-      }
-    }
-  }
-
-  // --- AI Studio: Translate (requires slug) ---
-  app.post("/api/studio/translate", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-
-    try {
-      const body = (await c.req.json().catch(() => ({}))) as {
-        providerId?: unknown;
-        modelId?: unknown;
-        toolName?: unknown;
-        targetLang?: unknown;
-        sourceLang?: unknown;
-      };
-      const ai = await resolveAiSelection(body, c.req.raw.signal);
-
-      const targetSession = await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const targetLang = typeof body.targetLang === "string" ? body.targetLang : "English";
-      const sourceLang = typeof body.sourceLang === "string" ? body.sourceLang : undefined;
-
-      // Load existing overlays BEFORE generation so we can chain operations
-      const existing = await loadOverlays(baseDir, result.slug, targetId);
-      // Remove translate overlays — we're replacing them. Keep others (tone etc.) for chaining.
-      const nonTranslateOverlays = existing.overlays.filter((o) => o.source.type !== "translate");
-      const chainBase: SessionOverlays = { version: 1, overlays: nonTranslateOverlays };
-      const effectiveSession = sessionWithEffectiveContent(targetSession, chainBase);
-
-      const translationResult = await generateTranslation(
-        effectiveSession,
-        ai.selection,
-        { targetLang, sourceLang },
-        { signal: c.req.raw.signal },
-      );
-      if (!translationResult) return c.json({ error: "Translation returned no result" }, 422);
-      // Restore true originalValue from the unmodified session
-      fixOriginalValues(translationResult.overlays, targetSession);
-      const merged: SessionOverlays = {
-        version: 1,
-        overlays: [...nonTranslateOverlays, ...translationResult.overlays],
-      };
-      await saveOverlays(baseDir, result.slug, merged, targetId);
-
-      return c.json({
-        overlays: merged,
-        stats: translationResult.stats,
-        providerId: ai.selection.providerId,
-        providerName: ai.providerName,
-        modelId: ai.modelId,
-        authType: ai.authType,
-        authSubscription: ai.authSubscription,
-        authSource: ai.authSource,
-      });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 500);
-    }
-  });
-
-  // --- AI Studio: Tone Adjustment (requires slug) ---
-  app.post("/api/studio/tone", async (c) => {
-    const result = requireSlug(c.req.query("slug"));
-    if ("error" in result) return c.json({ error: result.error }, 400);
-    const targetId = safeTargetId(c.req.query("targetId"));
-    if (targetId === null) return c.json({ error: "invalid targetId" }, 400);
-
-    try {
-      const body = (await c.req.json().catch(() => ({}))) as {
-        providerId?: unknown;
-        modelId?: unknown;
-        toolName?: unknown;
-        style?: unknown;
-      };
-      const ai = await resolveAiSelection(body, c.req.raw.signal);
-
-      const targetSession = await loadSessionFromDisk(baseDir, result.slug, targetId);
-      const style =
-        typeof body.style === "string" &&
-        ["professional", "neutral", "friendly"].includes(body.style)
-          ? (body.style as "professional" | "neutral" | "friendly")
-          : "professional";
-
-      // Load existing overlays BEFORE generation so we can chain operations
-      const existing = await loadOverlays(baseDir, result.slug, targetId);
-      // Remove tone overlays — we're replacing them. Keep others (translate etc.) for chaining.
-      const nonToneOverlays = existing.overlays.filter((o) => o.source.type !== "tone");
-      const chainBase: SessionOverlays = { version: 1, overlays: nonToneOverlays };
-      const effectiveSession = sessionWithEffectiveContent(targetSession, chainBase);
-
-      const toneResult = await generateToneAdjustment(
-        effectiveSession,
-        ai.selection,
-        { style },
-        { signal: c.req.raw.signal },
-      );
-      if (!toneResult) return c.json({ error: "Tone adjustment returned no result" }, 422);
-      // Restore true originalValue from the unmodified session
-      fixOriginalValues(toneResult.overlays, targetSession);
-      const merged: SessionOverlays = {
-        version: 1,
-        overlays: [...nonToneOverlays, ...toneResult.overlays],
-      };
-      await saveOverlays(baseDir, result.slug, merged, targetId);
-
-      return c.json({
-        overlays: merged,
-        stats: toneResult.stats,
-        providerId: ai.selection.providerId,
-        providerName: ai.providerName,
-        modelId: ai.modelId,
-        authType: ai.authType,
-        authSubscription: ai.authSubscription,
-        authSource: ai.authSource,
-      });
-    } catch (err) {
-      return c.json({ error: await getAiRuntime().getSafeErrorMessage(err) }, 500);
-    }
+  registerAiRoutes(app, {
+    baseDir,
+    loadSession: (slug, targetId) => loadSessionFromDisk(baseDir, slug, targetId),
+    isSameOriginSettingsRequest,
   });
 
   // Dev mode: use VIBE_API_PORT env (set by scripts/dev.mjs) or fall back to 13456
